Add auth library abstraction
