+++ src/login/AbstractLoginResult.ts
@@ -1,0 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { TokenResponse } from "adal-node";
+
+export type AbstractLoginResult = TokenResponse[];
