+++ src/login/AzureLoginHelper.ts
@@ -5,33 +5,25 @@
 
 import { SubscriptionClient, SubscriptionModels } from '@azure/arm-subscriptions';
 import { Environment } from '@azure/ms-rest-azure-env';
-import { DeviceTokenCredentials as DeviceTokenCredentials2, TokenCredentialsBase } from '@azure/ms-rest-nodeauth';
-import { Logging, MemoryCache, TokenResponse } from 'adal-node';
-import { DeviceTokenCredentials } from 'ms-rest-azure';
-import { CancellationTokenSource, commands, ConfigurationTarget, Disposable, EventEmitter, ExtensionContext, MessageItem, OutputChannel, QuickPickItem, window, workspace, WorkspaceConfiguration } from 'vscode';
+import { CancellationTokenSource, commands, ConfigurationTarget, Disposable, EventEmitter, ExtensionContext, MessageItem, QuickPickItem, window, workspace, WorkspaceConfiguration } from 'vscode';
 import { AzureAccount, AzureLoginStatus, AzureResourceFilter, AzureSession, AzureSubscription } from '../azure-account.api';
 import { createCloudConsole } from '../cloudConsole/cloudConsole';
-import { azureCustomCloud, azurePPE, cacheKey, clientId, cloudSetting, commonTenantId, customCloudArmUrlSetting, displayName, extensionPrefix, resourceFilterSetting, staticEnvironments, tenantSetting } from '../constants';
+import { azureCustomCloud, azurePPE, cacheKey, clientId, cloudSetting, commonTenantId, credentialsSection, customCloudArmUrlSetting, extensionPrefix, resourceFilterSetting, tenantSetting } from '../constants';
 import { AzureLoginError, getErrorMessage } from '../errors';
 import { TelemetryReporter } from '../telemetry';
 import { listAll } from '../utils/arrayUtils';
+import { KeyTar, tryGetKeyTar } from '../utils/keytar';
 import { localize } from '../utils/localize';
 import { openUri } from '../utils/openUri';
 import { getSettingValue, getSettingWithPrefix } from '../utils/settingUtils';
 import { delay } from '../utils/timeUtils';
+import { AbstractAuthProvider } from './AbstractAuthProvider';
+import { AbstractLoginResult } from './AbstractLoginResult';
 import { getEnvironments, getSelectedEnvironment, isADFS } from './environments';
 import { addFilter, getNewFilters, removeFilter } from './filters';
-import { login } from './login';
-import { loginWithDeviceCode } from './loginWithDeviceCode';
+import { getKey } from './getKey';
 import { checkRedirectServer } from './server';
-import { addTokenToCache, clearTokenCache, deleteRefreshToken, getStoredCredentials, getTokenWithAuthorizationCode, ProxyTokenCache, storeRefreshToken, tokenFromRefreshToken, tokensFromToken } from './tokens';
 import { waitUntilOnline } from './waitUntilOnline';
-
-const staticEnvironmentNames: string[] = [
-	...staticEnvironments.map(environment => environment.name),
-	azureCustomCloud,
-	azurePPE
-];
 
 const environmentLabels: Record<string, string> = {
 	AzureCloud: localize('azure-account.azureCloud', 'Azure'),
@@ -42,7 +34,8 @@ const environmentLabels: Record<string, string> = {
 	[azurePPE]: localize('azure-account.azurePPE', 'Azure PPE'),
 };
 
-const logVerbose: boolean = false;
+const enableVerboseLogs: boolean = false;
+const keytar: KeyTar | undefined = tryGetKeyTar();
 
 interface IAzureAccountWriteable extends AzureAccount {
 	status: AzureLoginStatus;
@@ -52,7 +45,7 @@ export interface ISubscriptionItem extends QuickPickItem {
 	subscription: AzureSubscription;
 }
 
-interface ISubscriptionCache {
+export interface ISubscriptionCache {
 	subscriptions: {
 		session: {
 			environment: string;
@@ -76,10 +69,10 @@ export class AzureLoginHelper {
 	private filtersTask: Promise<AzureResourceFilter[]> = Promise.resolve(<AzureResourceFilter[]>[]);
 	private onFiltersChanged: EventEmitter<void> = new EventEmitter<void>();
 
-	private tokenCache: MemoryCache = new MemoryCache();
-	private delayedTokenCache: ProxyTokenCache = new ProxyTokenCache(this.tokenCache);
 	private oldResourceFilter: string = '';
 	private doLogin: boolean = false;
+
+	private authProvider: AbstractAuthProvider;
 
 	public api: AzureAccount = {
 		status: 'Initializing',
@@ -97,6 +90,8 @@ export class AzureLoginHelper {
 	};
 
 	constructor(private context: ExtensionContext, private reporter: TelemetryReporter) {
+		this.authProvider = new AbstractAuthProvider(context, enableVerboseLogs);
+
 		context.subscriptions.push(commands.registerCommand('azure-account.login', () => this.login('login').catch(console.error)));
 		context.subscriptions.push(commands.registerCommand('azure-account.loginWithDeviceCode', () => this.login('loginWithDeviceCode').catch(console.error)));
 		context.subscriptions.push(commands.registerCommand('azure-account.logout', () => this.logout().catch(console.error)));
@@ -117,23 +112,6 @@ export class AzureLoginHelper {
 		}));
 		this.initialize('activation', false, true)
 			.catch(console.error);
-
-		if (logVerbose) {
-			const outputChannel: OutputChannel = window.createOutputChannel(displayName);
-			context.subscriptions.push(outputChannel);
-			Logging.setLoggingOptions({
-				level: 3 /* Logging.LOGGING_LEVEL.VERBOSE */,
-				// eslint-disable-next-line @typescript-eslint/no-explicit-any
-				log: (_level: any, message: any, error: any) => {
-					if (message) {
-						outputChannel.appendLine(message);
-					}
-					if (error) {
-						outputChannel.appendLine(error);
-					}
-				}
-			});
-		}
 	}
 
 	public async login(trigger: LoginTrigger): Promise<void> {
@@ -166,16 +144,10 @@ export class AzureLoginHelper {
 			const isAdfs: boolean = isADFS(environment);
 			const useCodeFlow: boolean = trigger !== 'loginWithDeviceCode' && await checkRedirectServer(isAdfs);
 			path = useCodeFlow ? 'newLoginCodeFlow' : 'newLoginDeviceCode';
-			const tokenResponse: TokenResponse = useCodeFlow ? 
-				await login(clientId, environment, isAdfs, tenantId, openUri, redirectTimeout) : 
-				await loginWithDeviceCode(environment, tenantId);
-
-			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-			const refreshToken: string = tokenResponse.refreshToken!;
-			const tokenResponses: TokenResponse[] = tenantId === commonTenantId ? await tokensFromToken(environment, tokenResponse) : [tokenResponse];
-
-			await storeRefreshToken(environment, refreshToken);
-			await this.updateSessions(environment, tokenResponses);
+			const loginResult: AbstractLoginResult = useCodeFlow ? 
+				await this.authProvider.login(clientId, environment, isAdfs, tenantId, openUri, redirectTimeout) :
+				await this.authProvider.loginWithDeviceCode(environment, tenantId);
+			await this.updateSessions(environment, loginResult);
 			void this.sendLoginTelemetry(trigger, path, environmentName, 'success', undefined, true);
 		} catch (err) {
 			if (err instanceof AzureLoginError && err.reason) {
@@ -217,11 +189,7 @@ export class AzureLoginHelper {
 
 	public async logout(): Promise<void> {
 		await this.api.waitForLogin();
-		// 'Azure' and 'AzureChina' are the old names for the 'AzureCloud' and 'AzureChinaCloud' environments
-		const allEnvironmentNames: string[] = staticEnvironmentNames.concat(['Azure', 'AzureChina', 'AzurePPE'])
-		for (const name of allEnvironmentNames) {
-			await deleteRefreshToken(name);
-		}
+		await this.authProvider.deleteRefreshTokens();
 		await this.clearSessions();
 		this.updateLoginStatus();
 	}
@@ -274,56 +242,19 @@ export class AzureLoginHelper {
 	private async initialize(trigger: LoginTrigger, doLogin?: boolean, migrateToken?: boolean): Promise<void> {
 		let environmentName: string = 'uninitialized';
 		try {
-			const showTimingLogs: boolean = false;
-			const start: number = Date.now();
 			await this.loadSubscriptionCache();
-			showTimingLogs && console.log(`loadSubscriptionCache: ${(Date.now() - start) / 1000}s`);
 			const environment: Environment = await getSelectedEnvironment();
 			environmentName = environment.name;
 			const tenantId: string = getSettingValue(tenantSetting) || commonTenantId;
 			const storedCreds: string | undefined = await getStoredCredentials(environment, migrateToken);
 
-			showTimingLogs && console.log(`keytar: ${(Date.now() - start) / 1000}s`);
 			if (!storedCreds) {
 				throw new AzureLoginError(localize('azure-account.refreshTokenMissing', "Not signed in"));
 			}
 			await waitUntilOnline(environment, 5000);
 			this.beginLoggingIn();
-
-			// eslint-disable-next-line @typescript-eslint/no-explicit-any
-			let parsedCreds: any;
-			let tokenResponse: TokenResponse | undefined;
-			try {
-				// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-				parsedCreds = JSON.parse(storedCreds);
-			} catch {
-				tokenResponse = await tokenFromRefreshToken(environment, storedCreds, tenantId);
-			}
-
-			if (parsedCreds) {
-				// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-				const { redirectionUrl, code } = parsedCreds;
-				if (!redirectionUrl || !code) {
-					throw new AzureLoginError(localize('azure-account.malformedCredentials', "Stored credentials are invalid"));
-				}
-
-				tokenResponse = await getTokenWithAuthorizationCode(clientId, Environment.AzureCloud, redirectionUrl, tenantId, code);
-			}
-
-			if (!tokenResponse) {
-				throw new AzureLoginError(localize('azure-account.missingTokenResponse', "Using stored credentials failed"));
-			}
-
-			showTimingLogs && console.log(`tokenFromRefreshToken: ${(Date.now() - start) / 1000}s`);
-			// For testing
-			if (workspace.getConfiguration(extensionPrefix).get('testTokenFailure')) {
-				throw new AzureLoginError(localize('azure-account.testingAcquiringTokenFailed', "Testing: Acquiring token failed"));
-			}
-
-			const tokenResponses: TokenResponse[] = tenantId === commonTenantId ? await tokensFromToken(environment, tokenResponse) : [tokenResponse];
-			showTimingLogs && console.log(`tokensFromToken: ${(Date.now() - start) / 1000}s`);
-			await this.updateSessions(environment, tokenResponses);
-			showTimingLogs && console.log(`updateSessions: ${(Date.now() - start) / 1000}s`);
+			const loginResult: AbstractLoginResult = await this.authProvider.loginSilent(environment, storedCreds, tenantId);
+			await this.updateSessions(environment, loginResult);
 			void this.sendLoginTelemetry(trigger, 'tryExisting', environmentName, 'success', undefined, true);
 		} catch (err) {
 			await this.clearSessions(); // clear out cached data
@@ -341,10 +272,10 @@ export class AzureLoginHelper {
 	}
 
 	private async loadSubscriptionCache(): Promise<void> {
-		const cache: ISubscriptionCache | undefined = this.context.globalState.get<ISubscriptionCache>(cacheKey);
+		const cache: ISubscriptionCache | undefined = this.context.globalState.get(cacheKey);
 		if (cache) {
 			(<IAzureAccountWriteable>this.api).status = 'LoggedIn';
-			const sessions: Record<string, AzureSession> = await this.initializeSessions(cache);
+			const sessions: Record<string, AzureSession> = await this.authProvider.initializeSessions(cache, this.api);
 			const subscriptions: AzureSubscription[] = this.initializeSubscriptions(cache, sessions);
 			this.initializeFilters(subscriptions);
 		}
@@ -383,52 +314,13 @@ export class AzureLoginHelper {
 		}
 	}
 
-	private async initializeSessions(cache: ISubscriptionCache): Promise<Record<string, AzureSession>> {
-		const sessions: Record<string, AzureSession> = {};
-		for (const { session } of cache.subscriptions) {
-			const { environment, userId, tenantId } = session;
-			const key: string = `${environment} ${userId} ${tenantId}`;
-			const environments: Environment[] = await getEnvironments();
-			const env: Environment | undefined = environments.find(e => e.name === environment);
-			if (!sessions[key] && env) {
-				sessions[key] = {
-					environment: env,
-					userId,
-					tenantId,
-					// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-explicit-any
-					credentials: new DeviceTokenCredentials({ environment: (<any>Environment)[environment], username: userId, clientId, tokenCache: this.delayedTokenCache, domain: tenantId }),
-					credentials2: new DeviceTokenCredentials2(clientId, tenantId, userId, undefined, env, this.delayedTokenCache)
-				};
-				this.api.sessions.push(sessions[key]);
-			}
-		}
-		return sessions;
-	}
-
-	private async updateSessions(environment: Environment, tokenResponses: TokenResponse[]): Promise<void> {
-		await clearTokenCache(this.tokenCache);
-		for (const tokenResponse of tokenResponses) {
-			await addTokenToCache(environment, this.tokenCache, tokenResponse);
-		}
-		/* eslint-disable @typescript-eslint/no-non-null-assertion */
-		this.delayedTokenCache.initEnd!();
-		const sessions: AzureSession[] = this.api.sessions;
-		sessions.splice(0, sessions.length, ...tokenResponses.map<AzureSession>(tokenResponse => ({
-			environment,
-			userId: tokenResponse.userId!,
-			tenantId: tokenResponse.tenantId!,
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-explicit-any
-			credentials: new DeviceTokenCredentials({ environment: (<any>environment), username: tokenResponse.userId, clientId, tokenCache: this.delayedTokenCache, domain: tokenResponse.tenantId }),
-			credentials2: new DeviceTokenCredentials2(clientId, tokenResponse.tenantId, tokenResponse.userId, undefined, environment, this.delayedTokenCache)
-		})));
+	private async updateSessions(environment: Environment, loginResult: AbstractLoginResult): Promise<void> {
+		await this.authProvider.updateSessions(environment, loginResult, this.api.sessions);
 		this.onSessionsChanged.fire();
-		/* eslint-enable @typescript-eslint/no-non-null-assertion */
 	}
 
 	private async clearSessions(): Promise<void> {
-		await clearTokenCache(this.tokenCache);
-		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-		this.delayedTokenCache.initEnd!();
+		await this.authProvider.clearTokenCache();
 		const sessions: AzureSession[] = this.api.sessions;
 		sessions.length = 0;
 		this.onSessionsChanged.fire();
@@ -445,7 +337,7 @@ export class AzureLoginHelper {
 	private initializeSubscriptions(cache: ISubscriptionCache, sessions: Record<string, AzureSession>): AzureSubscription[] {
 		const subscriptions: AzureSubscription[] = cache.subscriptions.map<AzureSubscription>(({ session, subscription }) => {
 			const { environment, userId, tenantId } = session;
-			const key: string = `${environment} ${userId} ${tenantId}`;
+			const key: string = getKey(environment, userId, tenantId);			
 			return {
 				session: sessions[key],
 				subscription
@@ -528,8 +420,7 @@ export class AzureLoginHelper {
 
 	private async loadSubscriptions(): Promise<AzureSubscription[]> {
 		const lists: AzureSubscription[][] = await Promise.all(this.api.sessions.map(session => {
-			const credentials: TokenCredentialsBase = session.credentials2;
-			const client: SubscriptionClient = new SubscriptionClient(credentials, { baseUri: session.environment.resourceManagerEndpointUrl });
+			const client: SubscriptionClient = new SubscriptionClient(session.credentials2, { baseUri: session.environment.resourceManagerEndpointUrl });
 			return listAll(client.subscriptions, client.subscriptions.list())
 				.then(list => list.map(subscription => ({
 					session,
@@ -634,4 +525,28 @@ function getCurrentTarget(config: { key: string; defaultValue?: unknown; globalV
 		}
 	}
 	return ConfigurationTarget.Global;
+}
+
+async function getStoredCredentials(environment: Environment, migrateToken?: boolean): Promise<string | undefined> {
+	if (!keytar) {
+		return undefined;
+	}
+	try {
+		if (migrateToken) {
+			const token = await keytar.getPassword('VSCode Public Azure', 'Refresh Token');
+			if (token) {
+				if (!await keytar.getPassword(credentialsSection, 'Azure')) {
+					await keytar.setPassword(credentialsSection, 'Azure', token);
+				}
+				await keytar.deletePassword('VSCode Public Azure', 'Refresh Token');
+			}
+		}
+	} catch {
+		// ignore
+	}
+	try {
+		return await keytar.getPassword(credentialsSection, environment.name) || undefined;
+	} catch {
+		// ignore
+	}
 }
